WireFrames

+++ wireframes/js/main.js
@@ -1,9 +1,16 @@
 var isEdited = null;
+var cancel = $('.change-row-cancel');
+var changeRow = $('.change-row');
+var stele = $('.stars');
+var myTable = $('#my-table');
+
 var onSubmit = function () {
     if (isEdited) {
         store.update(isEdited.id, getFormData()).then(function () {
             drawTable(store);
-
+            cancel.removeClass('display');
+            changeRow.removeClass('display');
+            isEdited = null;
         });
     } else {
         store.add(getFormData()).then(function () {
@@ -11,7 +18,7 @@ var onSubmit = function () {
         });
     }
     $('#form')[0].reset();
-    $('.stars').val("").change();
+    stele.val("").change();
     return false;
 };
 
@@ -23,18 +30,26 @@ function transform(bool) {
 var getFormData = function () {
     return {
         name: $('[name="city"]').val(),
-        stars: parseInt($('.stars').val()),
+        stars: parseInt(stele.val()),
         visited: transform($('[name="check"]').is(":checked"))
     };
 };
+//table.find('tr') != $('tr')
+// deseneaza tabel = drawForm()
 var page = 1;
 var totalPages = 1;
 
 var drawTable = function (store) {
-    store.getAll(page).then(function (data) {
+    store.getAll(page, sortField, sortDir).then(function (data) {
         $('tbody tr').remove();
 
         $.each(data.list, function () {
+            var currentStars = '';
+            for (i = 0; i <= this.stars - 1; i++) {
+                currentStars = currentStars + "★";
+            }
+            this.stars = currentStars;
+            this.visited = (this.visited == 0 ) ? "No" : "Yes";
             var tr = tmpl("item_tmpl", this);
             $('table tbody').append(tr);
             $('#pagins').text(data.page);
@@ -43,6 +58,7 @@ var drawTable = function (store) {
         totalPages = data.totalPages;
         $('#totalpagins').text(totalPages);
         attachEvents();
+
     });
 };
 
@@ -65,44 +81,76 @@ $('#prev').click(function () {
 
     return false;
 });
+var th = myTable.find('th');
+var sortField = th.data('fieldname');
+var sortDir = th.attr('data-sorttype');
+
+var sortUp = $('.sort-up');
+var sortDown = $('.sort-down');
+
+th.click(function () {
+    sortField = $(this).data('fieldname');
+
+    if (sortDir == 'asc') {
+        sortDir = 'desc';
+        sortDown.attr('id', 'display');
+        sortUp.attr('id', '');
+    } else {
+        sortDir = 'asc';
+        sortDown.attr('id', '');
+        sortUp.attr('id', 'display');
+    }
+    $(this).attr('data-sorttype', sortDir);
+
+    drawTable(store);
+});
 
 function attachEvents() {
     $('.remove').on('click', function () {
-        var id = $(this).closest('tr').data('id');
-        store.delete(id).then(function () {
-            drawTable(store);
-        });
+        if (confirm("Are you sure?")) {
+            var id = $(this).closest('tr').data('id');
+            store.delete(id).then(function () {
+                drawTable(store);
+            });
+        }
 
         return false;
     });
 
+    var checked = $('input[name="check"]');
     $('.edit').on('click', function () {
         var id = $(this).closest('tr').data('id');
         store.get(id).then(function (data) {
             isEdited = data;
             $('input[name="city"]').val(data.name);
-            $('.stars').val(data.stars).change();
+            stele.val(data.stars).change();
+
             if (data.visited == "1") {
-                $('input[name="check"]').prop("checked", true);
+                checked.prop("checked", true);
             } else {
-                $('input[name="check"]').prop("checked", false);
+                checked.prop("checked", false);
             }
         });
-        $('.change-row').addClass('display');
-        $('.change-row-cancel').addClass('display');
+        changeRow.addClass('display');
+        cancel.addClass('display');
     });
 }
 
-$('.change-row-cancel').click(function () {
+cancel.click(function () {
     $('#form')[0].reset();
-    $('.change-row-cancel').removeClass('display');
-    $('.change-row').removeClass('display');
-    $('.stars').val("").change();
+    cancel.removeClass('display');
+    changeRow.removeClass('display');
+    stele.val("").change();
     isEdited = null;
 
     return false;
 
 });
+
+$('.icon').click(function () {
+    $('.overlay').attr('id', '');
+});
+
 $(function () {
     var gif = $(".gif");
 
@@ -114,21 +162,46 @@ $(function () {
             gif.removeClass("display");
         }
     });
-    $('#form').not('.change-row-cancel').submit(onSubmit);
+
+    $('#form').submit(onSubmit);
     drawTable(store);
-
-});
-
-$(function () {
     $('[name="stea"]').stars({
         stars: 5
+    });
+
+    var tbody = $('tbody');
+    tbody.on('click', '.nume', function () {
+        var text = $(this).text();
+        $.ajax({
+            url: "http://api.giphy.com/v1/gifs/search?q=" + text + "&api_key=dc6zaTOxFJmzC&limit=1",
+            type: 'GET',
+            'Content-Type': 'application/json'
+        }).done(function (data) {
+            if (data.data.length == 0) {
+                alert('Sorry, no gif found');
+            } else {
+                $(".overlay").attr('id', 'display');
+                $.each(data.data, function (index) {
+                    var gifImage = data.data[index].images.original.url;
+                    $('.gif-here').html('<img src = "' + gifImage + '">');
+                });
+            }
+        });
     });
 });
 
 
+/*
+ q - search query term or phrase
+ limit - (optional) number of results to return, maximum 100. Default 25
+ offset - (optional) results offset, defaults to 0
+ rating - limit results to those rated (y,g, pg, pg-13 or r).
+ fmt - (optional) return results in html or json format (useful for viewing responses as GIFs to debug/test)
 
 
 
+
+ */
